Change function visibility of run() from public to external
Updated the visibility of the run() function in the TestWithdraw contract from public to external. This change improves gas efficiency by restricting the function to external calls only, as it is intended to be invoked by external contracts or accounts. 
No other functionality or logic in the contract has been altered.